add presentation layer

+++ src/main/java/by/gritsuk/dima/dao/impl/BetDAO.java
@@ -10,21 +10,21 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-public class BetDAO extends AbstractJdbcDao<Bet,Long> implements GenericDao<Bet,Long> {
+public class BetDAO extends AbstractJdbcDao<Bet,Integer> implements GenericDao<Bet,Integer> {
 
     @Override
     protected List<Bet> parseResultSet(ResultSet rs) throws SQLException {
         List<Bet>bets=new ArrayList<>();
         while(rs.next()){
             Bet bet=new Bet();
-            bet.setId(rs.getLong("id"));
+            bet.setId(rs.getInt("id"));
             bet.setMinValue(rs.getDouble("min_value"));
             Bet.CompetitionEvent event=new Bet.CompetitionEvent();
-            event.setId(rs.getLong("id"));
+            event.setId(rs.getInt("id"));
             event.setEvent(rs.getString("event_name"));
             event.setFactor(rs.getDouble("factor"));
             bet.setEvent(event);
-            bet.setCompetition_id(rs.getLong("competition_id"));
+            bet.setCompetition_id(rs.getInt("competition_id"));
             bets.add(bet);
         }
         return bets;
@@ -34,17 +34,17 @@ public class BetDAO extends AbstractJdbcDao<Bet,Long> implements GenericDao<Bet,
     protected void prepareStatementForInsert(PreparedStatement statement, Bet object) throws SQLException {
         int i=0;
         statement.setDouble(++i,object.getMinValue());
-        statement.setLong(++i,object.getCompetition_id());
-        statement.setLong(++i,object.getEvent().getId());
+        statement.setInt(++i,object.getCompetition_id());
+        statement.setInt(++i,object.getEvent().getId());
     }
 
     @Override
     protected void prepareStatementForUpdate(PreparedStatement statement, Bet object) throws SQLException {
         int i=0;
         statement.setDouble(++i,object.getMinValue());
-        statement.setLong(++i,object.getCompetition_id());
-        statement.setLong(++i,object.getEvent().getId());
-        statement.setLong(++i,object.getId());
+        statement.setInt(++i,object.getCompetition_id());
+        statement.setInt(++i,object.getEvent().getId());
+        statement.setInt(++i,object.getId());
     }
 
     @Override
